requirements.txt
pillow
python-dotenv
web3
fastapi
uvicorn[standard]